start moving clients into separate mod
